extract shift form
